refactor(audit): give each engine its own directory
ClickHouse machinery and its migration set sat at the package root, so a
second engine had nowhere symmetric to go. One directory per engine, holding
the client, the migration runner and the migrations. stores/ is unchanged: it
groups the AuditStore implementations, which is a different axis.
