+++ packages/audit/lib/index.ts
@@ -1,6 +1,6 @@
-export { auditClickhouseClient } from './clickhouse.js';
+export { auditClickhouseClient } from './clickhouse/clickhouse.js';
 export { AUDIT_EXPORT_MAX_ROWS, AuditClient, InvalidAuditCursorError } from './client.js';
-export { migrate } from './migrate.js';
+export { migrate } from './clickhouse/migrate.js';
 export { ClickhouseAuditStore } from './stores/clickhouse.js';
 export { NoopAuditStore } from './stores/noop.js';
 export { PubSubAuditWriter } from './stores/pubsub.js';
